Aesthetic Changes
Adjusted box shadow on pinned, mention popouts, and create guild this change is subtle and still retains a poppy look. I have finally skinned Better Discords emote tooltips and lastly I have unfortunately had to revert vertical margins on the user list as it was causing abnormal behavior when scrolling. For now all updates will be bug fixes (if any) and aesthetic changes such as this one until a new update is released that requires skinning